agentmon: create agent tasks for pending scans
